+++ backend/store/etcd/store_test.go
@@ -30,7 +30,7 @@ func testWithEtcd(t *testing.T, f func(store.Store)) {
 	f(s)
 }
 
-func testWithEtcdClient(t *testing.T, f func(store.Store, *clientv3.Client)) {
+func testWithEtcdStore(t *testing.T, f func(*Store)) {
 	e, cleanup := etcd.NewTestEtcd(t)
 	defer cleanup()
 
@@ -43,6 +43,21 @@ func testWithEtcdClient(t *testing.T, f func(store.Store, *clientv3.Client)) {
 	require.NoError(t, s.CreateNamespace(context.Background(), types.FixtureNamespace("default")))
 
 	f(s)
+}
+
+func testWithEtcdClient(t *testing.T, f func(store.Store, *clientv3.Client)) {
+	e, cleanup := etcd.NewTestEtcd(t)
+	defer cleanup()
+
+	client, err := e.NewClient()
+	require.NoError(t, err)
+
+	s := NewStore(client, e.Name())
+
+	// Mock a default namespace
+	require.NoError(t, s.CreateNamespace(context.Background(), types.FixtureNamespace("default")))
+
+	f(s, client)
 }
 
 type genericObject struct {